Refactor and rename key modules for consistency
Reorganized "general" modules, files, and exceptions to "generic" for better naming clarity and consistency. Removed obsolete tree structure builders, introducing updated implementations in "generic/structure.py". Adjusted imports across the codebase accordingly.
