+++ bin/generic.py
