+++ fairylandfuture/exceptions/generic.py
