+++ fairylandfuture/structures/generic/__init__.py
