+++ fairylandfuture/structures/generic/api.py
